implemented more relative positions
Everything in Motor A box is positioned relative to the upper left
coordinates of the Motor A box. Similarly, Motor B is positioned
relative to Motor B box, C to C box, and D to D box.
Motor Statuses are positioned relative to the Motor Status box, and
Sensor Statuses are positioned relative to the Sensor Status Box
(*except not in the Update Function yet)
Motor gang stuff is all positioned relatively in each of the three
boxes.
Battery level stuff is positioned relative to the battery box.