all dials hooked up
issues with the rate at which the infinite loop update function runs.
Probably going to incorporate phaser.game.time, so we can control
precisely when the updates will occur, like every 20 ms or every 50 ms
or something. About every 17 ms would be 60 hz which is as good as most
screens could display