main.js responsive behavior - frame movement works
Gangs are responsive b/w 4 columns and 3 columns in the dashboard.
Added a move(x1,y1,x2,y2) method to Frame prototype.
Phaser positions of graphics increment, so for example, rectangle.x =
newX actually means rectangle.x += newX. This works, but when moving
the frames back to the original positions, we just have to set
rectangle.x = 0 (it seems that you can't increment back to either the
original position or just the previous position—need to figure that
out). That new method handles the two cases described here